pyproject.toml
[project]
name = "whatscrapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.5",
    "mypy>=1.10.0",
    "types-requests>=2.32.0.20241016",
    "types-urllib3>=1.26.25.14",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["whatscrapping"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
ignore_missing_imports = true
